python/requirements.txt
pydantic>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
neo4j>=5.0,<6.0
httpx>=0.27,<1.0
